refactor(Events): seperate controller & service
